long constants end to end; SysV call parity now tracks live pushes
64-bit integer literals survive the whole pipeline: the lexer keeps the
full value (gcc's type ladder: decimal int->long->unsigned long, hex
int->unsigned->long->unsigned long, with the L/U suffixes), codegen
emits movabs imm64 for long constants, case labels and .quad static
slots carry the full 64-bit value (with the reg-indirect cmp when the
immediate outgrows the signed 32-bit form), sizeof/alignof emit as
%ld, the va_arg register offsets cast through (int) after the val
widen, and the (int) cast of an unsigned sign-extends for 32-bit
sources so (int)0xffffffffU is -1. the ~ right operand complements
32-bit operands in %eax so ~0U stays 0xffffffff.

call sites now pad to the real parity: a counter bumps on every push/
spill and pops on its matching pop, so a call issued with a save-push
still live from an enclosing store (memcpy filler, compound-assign
operand, index walk, ++/-- address) sees rsp 16-aligned exactly like
the plain odd-arg filler path (which folds into the same modulo).
every cluster pairs its pushes and pops inside straight-line paths, so
the counter stays branch-proof; it is reset per function entry.
cross-checked against gcc on generated stress programs (19 checks) and
the full selftest incl. the two-stage self-host cmp.

+++ src/ast.c
@@ -270,7 +270,7 @@ Type *func_type(Type *ret) {
   return t;
 }
 
-static CVal cv_int(int v)   { CVal c = {0, v, 0};    return c; }
+static CVal cv_int(long v) { CVal c = {0, v, 0}; return c; }
 static CVal cv_fp(double f) { CVal c = {1, 0, f};    return c; }
 
 /* every node kind const_fold() can evaluate, i.e. an integer
@@ -321,7 +321,7 @@ CVal const_fold(Node *n) {
         error("unsupported global initializer");
       {
         CVal c = const_fold(n->lhs);
-        return cv_int(c.is_float ? (int)c.fval : c.val);
+        return cv_int(c.is_float ? (long)c.fval : c.val);
       }
     case ND_UNARY:
       switch (n->op) {
@@ -373,8 +373,8 @@ CVal const_fold(Node *n) {
           default: error("unsupported global initializer");
         }
       } else {
-        int lv = l.val;
-        int rv = r.val;
+        long lv = l.val;
+        long rv = r.val;
         switch (n->op) {
           case '+': return cv_int(lv + rv);
           case '-': return cv_int(lv - rv);
